Created a separate script for the json generator class, added CBL version guardrails to multipeer tests
