+++ tests/QE/test_multipeer.py
@@ -434,6 +434,8 @@ class TestMultipeer(CBLTestClass):
 
     @pytest.mark.asyncio(loop_scope="session")
     async def test_scalable_conflict_resolution(self, cblpytest: CBLPyTest):
+        for ts in cblpytest.test_servers:
+            await self.skip_if_cbl_not(ts,">= 3.3.0")
         self.mark_test_step("Reset local database and load `empty` dataset on all devices")
 
         reset_tasks = [ts.create_and_reset_db(["db1"]) for ts in cblpytest.test_servers]
@@ -664,6 +666,8 @@ class TestMultipeer(CBLTestClass):
 
     @pytest.mark.asyncio(loop_scope="session")
     async def test_large_document_replication(self, cblpytest: CBLPyTest):
+        for ts in cblpytest.test_servers:
+            await self.skip_if_cbl_not(ts,">= 3.3.0")
         self.mark_test_step("Reset local database and load `empty` dataset on all devices")
 
         reset_tasks = [ts.create_and_reset_db(["db1"]) for ts in cblpytest.test_servers]
