+++ tests/QE/test_system_multipeer.py
@@ -25,132 +25,15 @@ from cbltest.api.replicator_types import (
     ReplicatorType,
 )
 from cbltest.api.test_functions import compare_local_and_remote
-
-
-class JSONGenerator:
-    def __init__(self, seed=random.randint(0, sys.maxsize), size=60000, format="json"):
-        self.seed = seed
-        self.size = size
-        self.format = format
-
-    def generate_document(self, doc_id: str) -> Dict[str, Any]:
-        """Generate a single JSON document with reproducible random data"""
-        random.seed(self.seed + int(doc_id.split("-")[0], 16))
-        if self.format == "json":
-            return {
-                doc_id: {
-                    "data": {
-                        "temperature": random.uniform(-20, 40),
-                        "humidity": random.randint(0, 100),
-                        "status": random.choice(["active", "inactive", "maintenance"]),
-                    },
-                    "metadata": {
-                        "version": 1,
-                        "created_at": int(time.time()),
-                        "modified_at": int(time.time()),
-                    },
-                }
-            }
-        else:
-            return {
-                doc_id: [{
-                    "data": {
-                        "temperature": random.uniform(-20, 40),
-                        "humidity": random.randint(0, 100),
-                        "status": random.choice(["active", "inactive", "maintenance"]),
-                    }},
-                    {"metadata": {
-                        "version": 1,
-                        "created_at": int(time.time()),
-                        "modified_at": int(time.time()),
-                    }}
-                ]
-            }
-
-    def update_document(self, doc: Any, doc_id: str) -> Dict[str, Any]:
-        """Update a document with reproducible modifications"""
-        offset = int(doc_id.split("-")[0], 16)
-        random.seed(self.seed + offset)
-        if self.format == "json":
-            doc["data"]["temperature"] += random.uniform(-5, 5)
-            doc["data"]["humidity"] = (doc["data"]["humidity"] + random.randint(-10, 10)) % 100
-            doc["data"]["status"] = random.choice(["active", "inactive", "maintenance"])
-            doc["metadata"]["version"] = doc["metadata"]["version"] + 1
-            doc["metadata"]["modified_at"] = int(time.time())
-
-        else:
-            doc[0]["data"] = {
-                "temperature": random.uniform(-20, 40),
-                "humidity": random.randint(0, 100),
-                "status": random.choice(["active", "inactive", "maintenance"]),
-            }
-            doc[1]["metadata"]["version"] = doc[1]["metadata"]["version"] + 1
-            doc[1]["metadata"]["modified_at"] = int(time.time())
-        return {doc_id: doc}
-
-    def batch_process(
-            self,
-            process_fn: Callable,
-            items_ids: List[Any],
-            items_doc: Dict[str, Any] = None,
-            batch_size: int = 1000
-    ) -> Dict[Any, Any]:
-        """Generic batch processing function with threading"""
-        results = {}
-
-        def process_batch(batch):
-            result = {}
-            for item in batch:
-                output = process_fn(items_doc[item], item) if items_doc is not None else process_fn(item)
-                result.update(output)
-            return result
-
-        with ThreadPoolExecutor() as executor:
-            futures = [
-                executor.submit(process_batch, items_ids[i:i + batch_size])
-                for i in range(0, len(items_ids), batch_size)
-            ]
-            for future in futures:
-                results.update(future.result())
-
-        return results
-
-    def generate_all_documents(self, size=None) -> Dict[str, Any]:
-        """Generate all documents using parallel processing"""
-        if size is None:
-            size = self.size
-        print(f"Generating {size} documents...")
-        start = time.time()
-
-        doc_ids = [str(uuid.uuid4()) for _ in range(size)]
-        documents = self.batch_process(
-            self.generate_document,
-            doc_ids
-        )
-
-        print(f"Generated {len(documents)} documents in {time.time() - start:.2f}s")
-        return documents
-
-    def update_all_documents(self, documents: Dict[str, Any]) -> Dict[str, Dict[str, Any]]:
-        """Update all documents with consistent modifications"""
-        print("Updating documents...")
-        start = time.time()
-
-        doc_ids = list(documents.keys())
-        updated = self.batch_process(
-            self.update_document,
-            doc_ids,
-            documents
-        )
-
-        print(f"Updated {len(updated)} documents in {time.time() - start:.2f}s")
-        return updated
+from cbltest.api.json_generator import JSONGenerator
 
 
 @pytest.mark.min_test_servers(2)
 class TestSystemMultipeer(CBLTestClass):
     @pytest.mark.asyncio(loop_scope="session")
     async def test_system(self, cblpytest: CBLPyTest):
+        for ts in cblpytest.test_servers:
+            await self.skip_if_cbl_not(ts,">= 3.3.0")
         NUM_DEVICES = len(cblpytest.test_servers)
         TEST_DURATION = timedelta(hours=1)
         CRUD_INTERVAL = timedelta(minutes=3)
@@ -301,6 +184,8 @@ class TestSystemMultipeer(CBLTestClass):
     # simultaneously inserting 10000 documents and blobs different docs to each CBL and starting replication
     @pytest.mark.asyncio(loop_scope="session")
     async def test_volume_with_blobs(self, cblpytest: CBLPyTest):
+        for ts in cblpytest.test_servers:
+            await self.skip_if_cbl_not(ts,">= 3.3.0")
         NO_OF_DOCS = 100000
         docgen = JSONGenerator(10, NO_OF_DOCS, format="key-value")
 
@@ -370,7 +255,8 @@ class TestSystemMultipeer(CBLTestClass):
         #   SGW1                    SGW2 ----  CBL5
         #    /                       \
         # CBL1 <--> CBL2<--->CBL3<--->CBL4
-
+        for ts in cblpytest.test_servers:
+            await self.skip_if_cbl_not(ts,">= 3.3.0")
         DOC_COUNT = 1000
 
         self.mark_test_step("Reset SG and load `names` dataset")
